+++ codyze-compliance/src/integrationTest/kotlin/de/fraunhofer/aisec/codyze/compliance/CommandTest.kt
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2025, Fraunhofer AISEC. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *                    $$$$$$\  $$$$$$$\   $$$$$$\
+ *                   $$  __$$\ $$  __$$\ $$  __$$\
+ *                   $$ /  \__|$$ |  $$ |$$ /  \__|
+ *                   $$ |      $$$$$$$  |$$ |$$$$\
+ *                   $$ |      $$  ____/ $$ |\_$$ |
+ *                   $$ |  $$\ $$ |      $$ |  $$ |
+ *                   \$$$$$   |$$ |      \$$$$$   |
+ *                    \______/ \__|       \______/
+ *
+ */
+package de.fraunhofer.aisec.codyze.compliance /*
+                                               * Copyright (c) 2025, Fraunhofer AISEC. All rights reserved.
+                                               *
+                                               * Licensed under the Apache License, Version 2.0 (the "License");
+                                               * you may not use this file except in compliance with the License.
+                                               * You may obtain a copy of the License at
+                                               *
+                                               *      http://www.apache.org/licenses/LICENSE-2.0
+                                               *
+                                               * Unless required by applicable law or agreed to in writing, software
+                                               * distributed under the License is distributed on an "AS IS" BASIS,
+                                               * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+                                               * See the License for the specific language governing permissions and
+                                               * limitations under the License.
+                                               *
+                                               *                    $$$$$$\  $$$$$$$\   $$$$$$\
+                                               *                   $$  __$$\ $$  __$$\ $$  __$$\
+                                               *                   $$ /  \__|$$ |  $$ |$$ /  \__|
+                                               *                   $$ |      $$$$$$$  |$$ |$$$$\
+                                               *                   $$ |      $$  ____/ $$ |\_$$ |
+                                               *                   $$ |  $$\ $$ |      $$ |  $$ |
+                                               *                   \$$$$$   |$$ |      \$$$$$   |
+                                               *                    \______/ \__|       \______/
+                                               *
+                                               */
+
+import com.github.ajalt.clikt.testing.test
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class CommandIntegrationTest {
+    @Test
+    fun testScanCommand() {
+        val command = ScanCommand()
+        val result =
+            command.test(
+                "--project-dir src/integrationTest/resources/demo-app --sources src/integrationTest/resources/demo-app/main.py"
+            )
+        assertEquals(
+            "Message(arguments=null, id=null, markdown=This is a **finding**, properties=null, text=null)\n",
+            result.output,
+        )
+    }
+}
