+++ codyze-compliance/src/test/kotlin/de/fraunhofer/aisec/codyze/compliance/CommandTest.kt
@@ -23,22 +23,25 @@
  *                    \______/ \__|       \______/
  *
  */
-import com.github.ajalt.clikt.testing.test
-import de.fraunhofer.aisec.cpg.codyze.compliance.ScanCommand
-import kotlin.test.Test
-import kotlin.test.assertEquals
+package de.fraunhofer.aisec.codyze.compliance
 
-class CommandIntegrationTest {
+import com.github.ajalt.clikt.testing.test
+import kotlin.test.*
+
+class CommandTest {
+
     @Test
-    fun testScanCommand() {
-        val command = ScanCommand()
-        val result =
-            command.test(
-                "--project-dir src/integrationTest/resources/demo-app --sources src/integrationTest/resources/demo-app/main.py"
-            )
-        assertEquals(
-            "Message(arguments=null, id=null, markdown=This is a **finding**, properties=null, text=null)\n",
-            result.output,
-        )
+    fun testComplianceCommand() {
+        val command = ComplianceCommand()
+        val result = command.test()
+        assertEquals(0, result.statusCode)
+    }
+
+    @Test
+    fun testListSecurityGoalsCommand() {
+        val command = ListSecurityGoals()
+        val result = command.test("--project-dir src/test/resources/")
+        assertEquals(0, result.statusCode)
+        assertEquals("Goal1\n", result.stdout)
     }
 }
