--- codyze-compliance/src/test/kotlin/de/fraunhofer/aisec/cpg/codyze/compliance/CommandTest.kt
@@ -1,47 +1,0 @@
-/*
- * Copyright (c) 2025, Fraunhofer AISEC. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *                    $$$$$$\  $$$$$$$\   $$$$$$\
- *                   $$  __$$\ $$  __$$\ $$  __$$\
- *                   $$ /  \__|$$ |  $$ |$$ /  \__|
- *                   $$ |      $$$$$$$  |$$ |$$$$\
- *                   $$ |      $$  ____/ $$ |\_$$ |
- *                   $$ |  $$\ $$ |      $$ |  $$ |
- *                   \$$$$$   |$$ |      \$$$$$   |
- *                    \______/ \__|       \______/
- *
- */
-package de.fraunhofer.aisec.cpg.codyze.compliance
-
-import com.github.ajalt.clikt.testing.test
-import kotlin.test.*
-
-class CommandTest {
-
-    @Test
-    fun testComplianceCommand() {
-        val command = ComplianceCommand()
-        val result = command.test()
-        assertEquals(0, result.statusCode)
-    }
-
-    @Test
-    fun testListSecurityGoalsCommand() {
-        val command = ListSecurityGoals()
-        val result = command.test("--project-dir src/test/resources/")
-        assertEquals(0, result.statusCode)
-        assertEquals("Goal1\n", result.stdout)
-    }
-}
